github-manager/requirements.txt
# Core dependencies
PyGithub>=2.1.1
GitPython>=3.1.40
PyYAML>=6.0.1
click>=8.1.7
aiofiles>=23.2.1
aiohttp>=3.9.1
websockets>=12.0

# Web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Utilities
python-dotenv>=1.0.0
tabulate>=0.9.0
pydantic>=2.5.0
httpx>=0.25.2

# Development tools
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Optional for enhanced features
rich>=13.7.0  # Better terminal output
tenacity>=8.2.3  # Retry logic
cachetools>=5.3.2  # Caching utilities